feat(app): add sensor creation modal
