+++ apps/app/app/admin/sensors/page.tsx
@@ -11,7 +11,7 @@ import { Row } from '@signalco/ui-primitives/Row';
 import { Stack } from '@signalco/ui-primitives/Stack';
 import { Typography } from '@signalco/ui-primitives/Typography';
 import { Suspense } from 'react';
-import { CreateSensorForm } from './CreateSensorForm';
+import { CreateSensorModal } from './CreateSensorModal';
 import { SensorServiceForm } from './SensorServiceForm';
 
 const statusLabels: Record<string, string> = {
@@ -134,10 +134,12 @@ export default async function SensorsPage() {
         0,
     );
 
-    const createFormRaisedBeds = raisedBeds.map((bed) => ({
-        id: bed.id,
-        physicalId: bed.physicalId,
-    }));
+    const createFormRaisedBeds = raisedBeds
+        .filter((bed) => bed.status === 'active')
+        .map((bed) => ({
+            id: bed.id,
+            physicalId: bed.physicalId,
+        }));
 
     return (
         <Stack spacing={2}>
@@ -146,8 +148,8 @@ export default async function SensorsPage() {
                     {'Senzori'}
                 </Typography>
                 <Chip color="primary">{totalSensors}</Chip>
+                <CreateSensorModal raisedBeds={createFormRaisedBeds} />
             </Row>
-            <CreateSensorForm raisedBeds={createFormRaisedBeds} />
             <Stack spacing={2}>
                 {sensorsByPhysicalId.length === 0 ? (
                     <Typography>Nema senzora.</Typography>
